+++ src/main/java/me/pajic/toolpouch/util/ClientUtil.java
@@ -2,6 +2,12 @@ package me.pajic.toolpouch.util;
 
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.Identifier;
+import net.minecraft.tags.ItemTags;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.LanternBlock;
+
+import java.util.function.Predicate;
 
 public class ClientUtil {
 
@@ -12,5 +18,9 @@ public class ClientUtil {
 		ToolPouchUtil.ITEM_SUGGESTIONS.clear();
 		ToolPouchUtil.ITEM_SUGGESTIONS.addAll(BuiltInRegistries.ITEM.keySet().stream().map(Identifier::toString).toList());
 		BuiltInRegistries.ITEM.listTagIds().forEach(tag -> ToolPouchUtil.ITEM_SUGGESTIONS.add("#" + tag.location()));
+	}
+
+	public static Predicate<ItemStack> getSupportedLanterns() {
+		return stack -> stack.is(ItemTags.LANTERNS) || Block.byItem(stack.getItem()) instanceof LanternBlock;
 	}
 }